+++ example/lib/main.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_osm_plugin/flutter_osm_plugin.dart';
 import 'package:flutter_osm_plugin_example/src/search_example.dart';
+import 'package:flutter_osm_plugin_example/src/simple_example_hook.dart';
 
 //import 'src/adv_home/home_example.dart';
 import 'src/home/home_example.dart';
@@ -28,10 +29,11 @@ class MyApp extends StatelessWidget {
       initialRoute: "/home",
       routes: {
         "/home": (ctx) => MainExample(),
+        "/simple": (ctx) => SimpleExample(),
         //"/adv-home": (ctx) => AdvandedMainExample(),
         // "/nav": (ctx) => MyHomeNavigationPage(
         //       map: Container(),
-        //     ),
+        // ),
         "/second": (ctx) => Scaffold(
               body: Center(
                 child: ElevatedButton(