+++ flutter_osm_interface/lib/src/types/bounding_box.dart
@@ -114,4 +114,8 @@ extension ExtBoundingBox on BoundingBox {
 
     return lonMatch && latMatch;
   }
+
+  List<double> toIOSList() {
+    return [south, west, north, east];
+  }
 }