+++ lib/src/controller/osm/osm_controller.dart
@@ -248,7 +248,7 @@ class MobileOSMController extends IBaseOSMController {
       initPosition = await myLocation();
       _checkBoundingBox(box, initPosition);
     }
-    if (box != null && !box.isWorld()) {
+    if (box != null && !box.isWorld() && Platform.isAndroid) {
       await limitAreaMap(box);
     }
     if (initPosition != null && !_osmFlutterState.setCache.value) {
@@ -313,6 +313,13 @@ class MobileOSMController extends IBaseOSMController {
       stepZoom,
     );
   }
+
+  @override
+  Future<void> changeTileLayer({CustomTile? tileLayer}) =>
+      osmPlatform.changeTileLayer(
+        _idMap,
+        tileLayer,
+      );
 
   /// set area camera limit of the map
   /// [box] : (BoundingBox) bounding that map cannot exceed from it
@@ -469,10 +476,13 @@ class MobileOSMController extends IBaseOSMController {
   }
 
   /// enabled tracking user location
-  Future<void> enableTracking() async {
+  Future<void> enableTracking({bool enableStopFollow = false}) async {
     /// make in native when is enabled ,nothing is happen
     await _osmFlutterState.requestPermission();
-    await osmPlatform.enableTracking(_idMap);
+    await osmPlatform.enableTracking(
+      _idMap,
+      stopFollowInDrag: enableStopFollow,
+    );
   }
 
   /// disabled tracking user location